requirements.txt
PySide6>=6.5.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
pysubs2>=1.6.0
psutil>=5.9.0
PySide6
